[FIX] search in booking_calendar
