+++ booking_calendar/models.py
@@ -118,12 +118,18 @@ class sale_order_line(models.Model):
             overlaps = 0
             if line.resource_id and line.booking_start and line.booking_end:
                 ids = getattr(self, '_origin', False) and self._origin.ids or bool(line.id) and [line.id] or []
-                overlaps = line.search_count(['&', '|', '&', ('booking_start', '>=', line.booking_start), ('booking_start', '<', line.booking_end),
-                                              '&', ('booking_end', '>', line.booking_start), ('booking_end', '<=', line.booking_end),
+                overlaps = line.search_count(['&', '|', '&', ('booking_start', '>', line.booking_start), ('booking_start', '<', line.booking_end),
+                                              '&', ('booking_end', '>', line.booking_start), ('booking_end', '<', line.booking_end),
                                               ('resource_id', '!=', False),
                                               ('id', 'not in', ids),
                                               ('resource_id', '=', line.resource_id.id),
                                               ('state', '!=', 'cancel')])
+                overlaps += line.search_count([('id', 'not in', ids),
+                                               ('booking_start', '=', line.booking_start),
+                                               ('booking_end', '=', line.booking_end),
+                                               ('resource_id', '=', line.resource_id.id),
+                                               ('state', '!=', 'cancel')])
+
             line.overlap = bool(overlaps)
 
     @api.multi
